# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi application",
        "version": "1.0"
    },
    "paths": {
...
        "/api/workers": {
            "get": {
                "tags": [
                    "Worker"
                ],
                "summary": "Get all workers",
                "operationId": "getAllWorkers",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Worker"
                ],
                "summary": "Post Worker",
                "operationId": "postWorker",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Ivan Ivanov",
                                "age": 33,
                                "company_id": 3
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Everything is fine"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/workers/{worker_id}": {
            "get": {
                "tags": [
                    "Worker"
                ],
                "summary": "Get Worker by id",
                "operationId": "getWorker",
                "parameters": [
                    {
                        "name": "worker_id",
                        "in": "path",
                        "description": "Worker ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Worker"
                ],
                "summary": "Update Worker by id",
                "operationId": "updateWorker",
                "parameters": [
                    {
                        "name": "worker_id",
                        "in": "path",
                        "description": "Worker ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Ivan Ivanov",
                                "age": 33,
                                "company_id": 3
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Worker"
                ],
                "summary": "Delete worker by id",
                "operationId": "deleteWorker",
                "parameters": [
                    {
                        "name": "worker_id",
                        "in": "path",
                        "description": "Worker ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Worker"
                ],
                "summary": "Change Worker by id",
                "operationId": "changeWorker",
                "parameters": [
                    {
                        "name": "worker_id",
                        "in": "path",
                        "description": "Worker ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Ivan Ivanov",
                                "age": 33,
                                "company_id": 3
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}